backend/requirements.txt
fastapi==0.111.0
mangum==0.17.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
boto3==1.34.84
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
